third_party/blink/web_tests/platform/generic/fast/forms/select/menulist-submit-without-selection-expected.txt

A unselected option was submitted as fallback. This behavior was removed by the change of webkit.org/b/35056.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS query.indexOf("select=Disabled") is -1
PASS successfullyParsed is true

TEST COMPLETE

Disabled
